Don't delete experiment folder if no results received yet
The folder is only created once a run completes, so if an in-progress
experiment is deleted it won't exist.

+++ weblab/experiments/signals.py
@@ -6,8 +6,11 @@ def experiment_version_deleted(sender, instance, **kwargs):
     Signal callback when an experiment version is about to be deleted.
 
     Ensure the experiment data directory is also deleted.
+
+    If the directory doesn't exist yet (because no results received) this is a no-op.
     """
-    rmtree(str(instance.abs_path))
+    if instance.abs_path.is_dir():
+        rmtree(str(instance.abs_path))
 
 
 def running_experiment_deleted(sender, instance, **kwargs):
